pdf_word_app/requirements.txt
Flask
python-docx
SQLAlchemy
PyPDF2
pdfplumber
